fix: renamed controller folder for failing build
